Lista de cambios: 1.- Cambio de iconos 2.- La camara del mapa, se utiliza para obtenr la posición de donde se quieren obtener los pokemons. 3.- Se añadio un pequeño servicio para poder mostrar de manera flotante el mapa.

+++ mobile/src/main/java/com/javic/pokewhere/models/Pokemon.java
@@ -45,7 +45,7 @@ public class Pokemon {
         this.expiration_time = expiration_time;
     }
 
-    public double getPokemonId() {
+    public long getPokemonId() {
         return pokemonId;
     }
